Add integration test for the publishers
We add everything to our previous test file and change the name
accordingly.
Also rename the test methods to guarantee the order of definition inside the
test class. We need to call the `test_lifecycle` last, because that
finalizes the driver from which it cannot recover.

There seem to be workarounds for keeping the order of unittests, e.g.
[here](https://stackoverflow.com/questions/4095319) or
[here](https://codereview.stackexchange.com/questions/122532),
but that seems less intuitive and a little overkill for the little number
integrations tests that we add and maintain here.